test and init connexion/inscription
