+++ pages/login.php
@@ -11,19 +11,15 @@
 
 <body>
     <?php require_once '../includes/navbar.php'; ?>
-
-    <!-- Message d'erreur général -->
     <?php if (!empty($error)) : ?>
         <p style="color: red;"><?php echo $error; ?></p>
     <?php endif; ?>
 
     <section class="containers_head_forms">
         <section class="container_forms">
-
             <h1>Bienvenue au monde, petit trésor !</h1>
             <p>Veuillez entrer vos informations pour vous connecter.</p>
-
-            <form action="inscription.php" method="POST">
+            <form action="Database.php" method="POST">
                 <label for="username">Nom d'utilisateur :</label>
                 <input type="text" id="username" name="username" placeholder="Entrez votre nom d'utilisateur" required><br><br>
 
@@ -32,7 +28,6 @@
 
                 <button type="submit">Se connecter</button>
             </form>
-
             <div class="message">
                 <?php if (isset($result['error'])): ?>
                     <p class="error"><?php echo $result['error']; ?></p>
